[Refactor] separate `renderer.inspector.ts` file into different files based on their responsibility
